Split module `hbranch` into package `hbranch` and introduce module `hbranch_v1`.
`hbranch` and `hbranch_v1` are now proper base and application models respectively.
