+++ hydpy/docs/hbranch_v1.rst
@@ -1,0 +1,9 @@
+
+.. _hbranch_v1:
+
+hbranch_v1 (HBV96 version of HydPy-H-Branch)
+============================================
+
+.. automodule:: hydpy.models.hbranch_v1
+    :members:
+    :show-inheritance:
